+++ dingtalk/java/src/main/java/com/aliyun/dingtalkworkbench_1_0/models/QueryShortcutScopesResponseBody.java
@@ -9,7 +9,7 @@ public class QueryShortcutScopesResponseBody extends TeaModel {
     public java.util.List<String> userVisibleScopes;
 
     @NameInMap("deptVisibleScopes")
-    public java.util.List<Float> deptVisibleScopes;
+    public java.util.List<Long> deptVisibleScopes;
 
     public static QueryShortcutScopesResponseBody build(java.util.Map<String, ?> map) throws Exception {
         QueryShortcutScopesResponseBody self = new QueryShortcutScopesResponseBody();
@@ -24,11 +24,11 @@ public class QueryShortcutScopesResponseBody extends TeaModel {
         return this.userVisibleScopes;
     }
 
-    public QueryShortcutScopesResponseBody setDeptVisibleScopes(java.util.List<Float> deptVisibleScopes) {
+    public QueryShortcutScopesResponseBody setDeptVisibleScopes(java.util.List<Long> deptVisibleScopes) {
         this.deptVisibleScopes = deptVisibleScopes;
         return this;
     }
-    public java.util.List<Float> getDeptVisibleScopes() {
+    public java.util.List<Long> getDeptVisibleScopes() {
         return this.deptVisibleScopes;
     }
